+++ src/antity-pgsql.d.ts
@@ -51,14 +51,24 @@ declare class SQLEntity extends Entity {
   get table(): string;
   set table(table: string);
   query: {
-    select: (paginate: boolean) => string;
-    update: (rows: Record<string, unknown>[], consumerId: number | string, consumerName: string) => {
-      query: string;
-      args: unknown[];
+    select: (
+      paginate: boolean,
+      first?: number,
+      rows?: number | null,
+      sortField?: string | null,
+      sortOrder?: "ASC" | "DESC" | null,
+      filters?: Filters | null) => {
+        query: string;
+        args: (Filter["value"])[];
+    };
+    update: (rows: Record<string,
+      unknown>[], consumerId: number | string, consumerName: string) => {
+        query: string;
+        args: unknown[];
     };
     insert: (rows: Record<string, unknown>[], consumerId: number | string, consumerName: string, rtn?: string) => {
-      query: string;
-      args: unknown[];
+        query: string;
+        args: unknown[];
     };
     delete: () => string;
     return: (prop: string) => string;